Se agregan los permisos de usuario para poder limitar las apciones

+++ noteBook/noteBook/UNA/vistas/LibroControlForm.cs
@@ -72,29 +72,46 @@ namespace noteBook.UNA.vistas
 
         private void EliminarBtn_Click(object sender, EventArgs e)
         {
-            MessageBoxButtons botones = MessageBoxButtons.YesNo;
-            DialogResult dr = MessageBox.Show("Seguro que desea eliminar el libro", "Alerta", botones, MessageBoxIcon.Warning);
-            if (dr == DialogResult.Yes)
+            MySqlDb mySqlDb = new MySqlDb();
+            mySqlDb.ConnectionString = System.Configuration.ConfigurationManager.ConnectionStrings["MySqlConnection"].ConnectionString;
+            mySqlDb.OpenConnection();
+            try
             {
-                if (dr == DialogResult.Yes)
+                string queryU = string.Format("Select id_usuario from usuarios where avatar='" + Singlenton.Instance.UsuarioActivo() + "'");
+
+                String queryPermiso = String.Format("Select id_permiso from permisos_personas where id_usuario='{0}'and id_permiso=3", mySqlDb.QuerySQL(queryU).Rows[0][0].ToString());
+                if (mySqlDb.QuerySQL(queryPermiso).Rows[0][0].ToString() == "3")
                 {
+                    MessageBoxButtons botones = MessageBoxButtons.YesNo;
+                    DialogResult dr = MessageBox.Show("Seguro que desea eliminar el libro se eliminaran las notas relacionadas", "Alerta", botones, MessageBoxIcon.Warning);
+                    if (dr == DialogResult.Yes)
+                    {
+                        if (dr == DialogResult.Yes)
+                        {
 
-                    MySqlDb mySqlDb = new MySqlDb();
-                    mySqlDb.ConnectionString = System.Configuration.ConfigurationManager.ConnectionStrings["MySqlConnection"].ConnectionString;
-                    mySqlDb.OpenConnection();
-                    String queryEliminarNota = String.Format("Delete from notas where id_libro=(select id_libro from libros where nombre='{0}')",this.nombre);
-                    String queryEliminarLibro = String.Format("Delete from  libros where nombre='{0}'",this.nombre);
-               
-                    mySqlDb.EjectSQL(queryEliminarNota);
-                    mySqlDb.EjectSQL(queryEliminarLibro);
-                    Singlenton.Instance.miLibro.CrearLibroDB();
+
+
+
+                            String queryEliminarNota = String.Format("Delete from notas where id_libro=(select id_libro from libros where nombre='{0}')", this.nombre);
+                            String queryEliminarLibro = String.Format("Delete from  libros where nombre='{0}'", this.nombre);
+
+                            mySqlDb.EjectSQL(queryEliminarNota);
+                            mySqlDb.EjectSQL(queryEliminarLibro);
+                            Singlenton.Instance.miLibro.CrearLibroDB();
+                        }
+                        ///TODOSinglenton.Instance.CargarReporte("Libro eliminado", $"Se elimino la libro {this.TituloLabel.Text}", $"Libro {this.TituloLabel.Text}");
+
+                    }
                 }
-                ///TODOSinglenton.Instance.CargarReporte("Libro eliminado", $"Se elimino la libro {this.TituloLabel.Text}", $"Libro {this.TituloLabel.Text}");
-
             }
+            catch (Exception Ex)
+            {
+                MessageBox.Show($"El usuario no tiene permiso para eliminar libros");
+            }
+
         }
 
-        
+
 
     }
 }
